Revamp the filters system
searchFilter was moved into filters as searchString.

filters is the new state object in imagesStore which consists of "basic" and "advanced" filters.

Basic filters orders by ascending or descending order, and filters by name or id ("date" in the ui).

Advanced filters, filters by resolution (exact,more than and less than) and image formats.

Filters reset on page reload, and on searchString reset, the current page of the gallery is set to 1.

Filters component takes care of updating the basic portion of the filters object, and advancedFiltersModal takes care of updating the advancedFilters part.

Both make use of the setFilters function from the useImages hook.

The imagesArray is first sorted, and then filtered.

The calculation of filteredImages contains a somewhat convoluted succesion of conditionals, this is intentional to iterate as few times as possible through sortedImages.