Massive rewrite of frontend wip
- implement monitor selection logic (wip).
- Add monitors modal.
- Add showMonitorModalOnStartup setting on app settings.
- Fix readActivePlaylist bug.
- Move first set of activePlaylist on first render to custom hook.
- Move stores to a stores directory.
- add AppConfiguration store.
- Update migration sql script.
- Update schema.
- Refine types.
- Delete app intro animation entirely (this was a workaround and is no
  longer needed).
- Implement shortcuts (not finished).
- Fix ugly first render (related to deleting app animation).
- Improve css and layout responsiveness.
- Move imagesStore from using context to zustand.

This commit got out of hand, it was supposed to be just the monitors
modals but one thing led to another and ended up touching a thousand
items in the list, plus I was coming up with stuff on the fly. I'll try
to make more atomic commits in the future.