Add back random image feature
This commit brings back the random image feature both for individual
playlists that support it and for the individual command in general,
it also moves the always start on first image setting from the app page
to the playlistConfigurationModal as it makes more sense to have it
there and also removed the swww animations toggle from the app menu and
is now controlled per playlist, or by setting none as the transition
type in the swww config file.

- Fix not removing playlist refs after stopping it.
- Add random image monitor configuration in app settings.
- Fix playlistConfigurationModal settings for timeofday and dayofweek
playlists.
- Fix sending stop playlist command of inexisting playlist.