Small daemon fixes
- Fix daemon duplicating activePlaylists on first startup, this was
caused because the daemon did not distinguish between a start playlist
command from the main electron process and when it was reading an old
activePlaylist from the db.
-Fix daemon running migration function on each connection to the
database.
-Fix bug in setImage retry logic in daemon playlist, while loop
condition was wrong.
-Fix kill daemon on exit not working correctly, this is because daemon
on stop sends a stop-daemon message back and other playlistController
instances had listeners for that message that restarted the daemon
process, this was fixed by creating a new connection without any
listeners on data to send the stop-daemon command.
-Fix daemon stopping all playlists on process exit, this was incorrect
behaviour, moved that logic to a new message action switch called
stop-all-playlists not yet implemented.