Add tray commands back and fix major bug.
- Add tray commands for controlling active playlists.
- Fix stopping a playlist from the tray not clearing the playlist in the
renderer.
- Fix elements being selected in modals when selecting images in the
gallery.
- Update tray to only show playlist controlls when there are active
playlists.
- Add notifications for pause and resume events in the daemon.
- Fix major bug caused by drizzle not adding the time property on images
that had a null value on the time column in the imagesInPlaylist table.

The bug was caused because there was an if statement when adding a new
playlist to the store, that checked if the image.time property wasn't
null, because drizzle did not even set the time property, this was
evaluating to undefined !== null, which is true, so undefined was being
set in the imagesTimeSet, when calculating the initialTimestamp for a
new image being added to the playlist,the first one would get a NaN and
pass, and NaN was added to the imagesTimeSet, then when the second one
was added, the while loop that checked if the current initialTimeStamp
was in the set always evaluated to true, because NaN was already in the
set, so the whole renderer proccesss got stuck in an infinite loop. This
was fixed by calculating a new image time if the image.time evaluates to
undefined or null, also this fixes another bug that ocurred when some
images had a null value, and thus when changing from other types of
playlists to a time of day playlist, the time inputs would display
empty, now they correctly timestamps.